update resource to work with if-match/if-none-match and etag header

+++ src/contact/resource/ContactResource.java
@@ -4,14 +4,20 @@ import java.util.List;
 
 import javax.inject.Singleton;
 import javax.ws.rs.*;
+import javax.ws.rs.core.CacheControl;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.EntityTag;
+import javax.ws.rs.core.HttpHeaders;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.MultivaluedMap;
 import javax.ws.rs.core.Request;
 import javax.ws.rs.core.Response;
+import javax.ws.rs.core.Response.ResponseBuilder;
 import javax.ws.rs.core.Response.Status;
 import javax.ws.rs.core.UriInfo;
 import javax.xml.bind.JAXBElement;
+
+import org.eclipse.jetty.http.HttpHeader;
 
 import contact.entity.Contact;
 import contact.entity.ContactList;
@@ -22,6 +28,7 @@ import contact.service.mem.MemDaoFactory;
 /**
  * Provide Contact web resource that response to HTTP request
  * With GET, GET with parameter, GET with query, POST, PUT with parameter, DELETE method.
+ * now GET with id parameter, POST and PUT support ETag Header.
  * @author Atit Leelasuksan 5510546221
  *
  */
@@ -34,6 +41,9 @@ public class ContactResource {
 	 * use to get data, save data and update data.
 	 */
 	private ContactDao dao;
+	private CacheControl cc;
+	private EntityTag eTag;
+	ResponseBuilder rb;
 	
 	/**
 	 * Initialize Resource and Contact Data Access Object.
@@ -41,6 +51,8 @@ public class ContactResource {
 	public ContactResource() {
 		DaoFactory.setFactory(new MemDaoFactory());
 		dao = DaoFactory.getInstance().getContactDao();
+		cc = new CacheControl();
+		cc.setMaxAge(3600);
 	}
 	
 	/**
@@ -52,7 +64,6 @@ public class ContactResource {
 	@GET
 	@Produces({MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON}) 
 	public Response getContact(@QueryParam("title") String searchText) {
-		System.out.println("---GET---");
 		if(searchText==null) {
 			ContactList contacts = new ContactList(dao.findAll());
 			if(contacts.getContactList().size()>0)
@@ -76,11 +87,16 @@ public class ContactResource {
 	@GET
 	@Path("{id}")
 	@Produces({MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON})
-	public Response getContactByID(@PathParam("id") long id) {
-		System.out.println("---GETID---");
+	public Response getContactByID(@PathParam("id") long id, @Context Request req) {
 		Contact contact = dao.find(id);
-		if(contact!=null)
-			return Response.ok(contact).build();
+		if(contact!=null) {
+			eTag = new EntityTag(contact.getLastUpdate().hashCode()+"");
+			rb = req.evaluatePreconditions(eTag);
+			if(rb!=null) {
+				return rb.status(Status.NOT_MODIFIED).cacheControl(cc).tag(eTag).build();
+			}
+			return Response.ok(contact).cacheControl(cc).tag(eTag).build();
+		}
 		return Response.noContent().build();
 	}
 	
@@ -94,14 +110,14 @@ public class ContactResource {
 	 */
 	@POST
 	@Consumes({MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON})
-	public Response createContactXML(JAXBElement<Contact> element, @Context UriInfo uriInfo) {
-		System.out.println("---POST---");
+	public Response createContactXML(JAXBElement<Contact> element, @Context UriInfo uriInfo, @Context Request req) {
 		Contact contact = element.getValue();
 		if(dao.find(contact.getId())==null) {
 			boolean isSuccess = dao.save(contact);
 			if( isSuccess ) {
 				System.out.println("Create name:" + contact.getName());
-				return Response.created(uriInfo.getBaseUriBuilder().path("/contacts/{id}").build(contact.getId())).build();
+				eTag = new EntityTag(contact.getLastUpdate().hashCode()+"");
+				return Response.created(uriInfo.getBaseUriBuilder().path("/contacts/{id}").build(contact.getId())).cacheControl(cc).tag(eTag).build();
 			}
 			System.out.println("CREATE ERROR");
 			return Response.status(Status.BAD_REQUEST).build();
@@ -123,15 +139,18 @@ public class ContactResource {
 	@PUT
 	@Path("{id}")
 	@Consumes({MediaType.APPLICATION_XML, MediaType.APPLICATION_JSON})
-	public Response updateContact(@PathParam("id") long id, JAXBElement<Contact> element, @Context UriInfo uriInfo) {
-		System.out.println("---PUT---");
+	public Response updateContact(@PathParam("id") long id, JAXBElement<Contact> element, @Context UriInfo uriInfo, @Context Request req) {
 		Contact contact = element.getValue();
+		Contact testCon = dao.find(id);
+		eTag = new EntityTag(testCon.getLastUpdate().hashCode()+"");
+		rb = req.evaluatePreconditions(eTag);
+		if(rb==null) return Response.status(Status.PRECONDITION_FAILED).build();
 		if(contact!=null) {
 			contact.setId(id);
 			boolean isSuccess = dao.update(contact);
 			if( isSuccess ) {
 				System.out.println("Update id:"+id);
-				return Response.ok().build();
+				return Response.ok().cacheControl(cc).tag(eTag).build();
 			}
 		}
 		System.out.println("UPDATE ERROR");
@@ -147,12 +166,17 @@ public class ContactResource {
 	 */
 	@DELETE
 	@Path("{id}")
-	public Response deleteContact(@PathParam("id") long id) {
-		System.out.println("---DELETE---");
-		boolean isSuccess = dao.delete(id);
-		if( isSuccess ) {
-			System.out.println("Delete id:" + id);
-			return Response.ok().build();
+	public Response deleteContact(@PathParam("id") long id, @Context Request req) {
+		Contact testCon = dao.find(id);
+		if(testCon!=null) {
+			eTag = new EntityTag(testCon.getLastUpdate().hashCode()+"");
+			rb = req.evaluatePreconditions(eTag);
+			if(rb==null) return Response.status(Status.PRECONDITION_FAILED).build();
+			boolean isSuccess = dao.delete(id);
+			if( isSuccess ) {
+				System.out.println("Delete id:" + id);
+				return Response.ok().build();
+			}
 		}
 		System.out.println("DELETE ERROR");
 		return Response.status(Status.BAD_REQUEST).build();
